Change the nomeclature for portfolios to groups
There are several customers that do not use the concept of portfolios.
Therefore it would not make sense for them to utilise portfolios in
their API calls. We instead introduce the concept of "groups" which
serve the same purpose as portfolios and internally utilise the
portfolio functionality.

Add an anchor for the glossary in the top-left corner of the navigation.