Renamed vacancy to listing and added group access errors
Vacancy needs to be referred in public facing documentation as listing
so any reference to vacancy has been changed to listing.

Added error messages and write up to relevant endpoints for informing
users of API behaviour if a group that they don't have access to is
provided in the API request.