Various updates to the docs structure
Changed the navigation structure so that everything is top-level instead
of fenced in arbitrary sub-groups.

Changed the text on the introduction page to include the text previously
present on the authentication page, as it included a call to action and a
link for siging up to the API.

Removed the call-to-action and sign up text from the authentication page.