Remove redundant endpoints
Remove endpoints that allow retrieval of a single record in favour of
endpoints that return multiple records. If a user wants to retrieve a
single record, they can narrow the search on the multiple record
retrieval endpoint.

Remove endpoints for creating single records. Users can make use of the
multi-record creation endpoints to create single users by passing
information for a single user.