Add an endpoint for syncing listings
The industry standard way for managing listings involve synchronising
a listing database with a provided list of listings periodically.

The sync endpoint will provide this functionality, as it allows listings
to be created or updated from a list, but also automatically deactivates/
deletes listings that are not in the provided list.