Updates to portfolio functionality and to vacancy find endpoint
We are allowing users to pass in their portfolio external references
to scope requests instead of only supporting internal portfolio IDs.

Only supporting internal portfolio IDs would mean that API users would
have to maintain a mapping between their own identifiers for their
corresponding portfolio entity and our own internal portfolio IDs.
We will be doing that for them in the Travtus API as we already have
the information that we need and it would take minimal development
effort to implement this.

Modify the find vacancy endpoint to reflect the fact that vacancies
now have external references.