Add errors to listing endpoints and fix bugs with request examples on several endpoints
Add errors for the listing find and sync endpoints for cases where a user attempts to
perform one of these operations on a portfolio that they don't have access to or when
a user does not provide a portfolio id or external reference when they have to provide
one.

Fix a bug on several endpoints where missing closing single and double quotes broke
request and response examples.