Specify the expected date format and add extra error for sync endpoint
API users need to only specify a year, month and date for date fields. Add
extra documentation for these fields to make it clear.

Add an extra error for the sync endpoint as it requires for an external
reference to be specified at the moment for all vacancies to be synced.